import inventory from revstore fun. upd

+++ org_data/008_pre_inventory.sql
@@ -22,7 +22,7 @@ declare
     cat3_id int;
     cat4_id int;
     unit_id int;
-
+    gst_tax text := 'gst0';
 begin
 
     begin
@@ -32,22 +32,26 @@ begin
       select id into cat3_id from category_option where name = new.category3 and category_id='INV_CAT3';
       select id into cat4_id from category_option where name = new.category4 and category_id='INV_CAT4';
       select id into unit_id from unit where name = 'PCS';
-
-      insert into inventory(name, division_id, gst_tax_id, unit_id, sale_unit_id, purchase_unit_id, hsn_code,
-                            category1, category2, category3, category4)
-                  values(new.name, div_id, new.gst_tax, unit_id,unit_id,unit_id,
-                        (case when length(trim(new.hsn_code))>0 then trim(new.hsn_code) else null end),
-                        (case when cat1_id is null then null else ARRAY[cat1_id]::int[] end),
-                        (case when cat2_id is null then null else ARRAY[cat2_id]::int[] end),
-                        (case when cat3_id is null then null else ARRAY[cat3_id]::int[] end),
-                        (case when cat4_id is null then null else ARRAY[cat4_id]::int[] end));
-      delete from temp_inv where id=new.id;
+      if (new.gst_tax <> 'gst') then
+          gst_tax = new.gst_tax;
+      end if;
+      if (new.name is not null and char_length(new.name) > 0) then
+          insert into inventory(name, division_id, gst_tax_id, unit_id, sale_unit_id, purchase_unit_id, hsn_code,
+                                category1, category2, category3, category4)
+                      values(new.name, div_id, gst_tax, unit_id, unit_id, unit_id,
+                            (case when (new.hsn_code ~ '^[0-9]*$') and (char_length(new.hsn_code) between 1 and 10) then trim(new.hsn_code) else null end),
+                            (case when cat1_id is null then null else ARRAY[cat1_id]::int[] end),
+                            (case when cat2_id is null then null else ARRAY[cat2_id]::int[] end),
+                            (case when cat3_id is null then null else ARRAY[cat3_id]::int[] end),
+                            (case when cat4_id is null then null else ARRAY[cat4_id]::int[] end));
+          delete from temp_inv where id=new.id;
+      end if;
 
     exception when others then
       raise exception 'insert into inventory(name, division_id, gst_tax_id, unit_id, sale_unit_id, purchase_unit_id, hsn_code,
                             category1, category2, category3, category4)
                   values(%,%,%,%,%,%,%,%,%,%,%);',new.name, div_id, new.gst_tax, unit_id,unit_id,unit_id,
-                        (case when length(trim(new.hsn_code))>0 then trim(new.hsn_code) else null end),
+                        (case when (new.hsn_code ~ '^[0-9]*$') and (char_length(new.hsn_code) between 1 and 10) then trim(new.hsn_code) else null end),
                         (case when cat1_id is null then null else ARRAY[cat1_id]::int[] end),
                         (case when cat2_id is null then null else ARRAY[cat2_id]::int[] end),
                         (case when cat3_id is null then null else ARRAY[cat3_id]::int[] end),
